Move calibration helper functions to 'util' submodule.
